moved buzzword card to its own component, retained functionality
